[Feature/sign-up] Complete html and initial CSS(excluding interactive states). Also add route mapping for login and signup page in RouteController.
