+++ src/main/java/com/decormasters/homedecor/controller/routes/RouteController.java
@@ -6,6 +6,15 @@ import org.springframework.web.bind.annotation.GetMapping;
 @Controller
 public class RouteController {
 
+    @GetMapping("/sign-up")
+    public String signUp() {
+        return "auth/signUp";
+    }
+
+    @GetMapping("/login")
+    public String login() {
+        return "auth/login";
+    }
 }
