+++ src/main/webapp/WEB-INF/views/index.jsp
@@ -9,8 +9,6 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Homedecor Clone</title>
-  <%-- font-awesome--%>
-  <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css" rel="stylesheet">
 
   <!-- Modular CSS -->
   <link rel="stylesheet" href="/css/main.css">
